+++ components/angular/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.3](https://github.com/matteobruni/tsparticles/compare/ng-particles@2.16.2...ng-particles@2.16.3) (2021-08-10)
+
+
+### Bug Fixes
+
+* fixed click handler, it wasn't working fine ([fd9873b](https://github.com/matteobruni/tsparticles/commit/fd9873b250d236b196cc7ff952d237f2768fb26b))
+
+
+### Features
+
+* added new methods to particle class ([5743453](https://github.com/matteobruni/tsparticles/commit/5743453906001569f262888aa54539ad4e1463ac))
+
+
+
+
+
 ## [2.16.2](https://github.com/matteobruni/tsparticles/compare/ng-particles@2.16.1...ng-particles@2.16.2) (2021-07-31)
 
 **Note:** Version bump only for package ng-particles