+++ components/riot/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/matteobruni/tsparticles/compare/riot-particles@1.7.2...riot-particles@1.7.3) (2021-08-10)
+
+
+### Features
+
+* added new methods to particle class ([5743453](https://github.com/matteobruni/tsparticles/commit/5743453906001569f262888aa54539ad4e1463ac))
+
+
+
+
+
 ## [1.7.2](https://github.com/matteobruni/tsparticles/compare/riot-particles@1.7.1...riot-particles@1.7.2) (2021-07-31)
 
 **Note:** Version bump only for package riot-particles