+++ demo/imports/ts34/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.3](https://github.com/matteobruni/tsparticles/compare/imports-ts34@1.4.2...imports-ts34@1.4.3) (2021-08-10)
+
+**Note:** Version bump only for package imports-ts34
+
+
+
+
+
 ## [1.4.2](https://github.com/matteobruni/tsparticles/compare/imports-ts34@1.4.1...imports-ts34@1.4.2) (2021-07-31)
 
 **Note:** Version bump only for package imports-ts34