+++ demo/react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.3](https://github.com/matteobruni/tsparticles/compare/react-tsparticles-demo@1.18.2...react-tsparticles-demo@1.18.3) (2021-08-10)
+
+**Note:** Version bump only for package react-tsparticles-demo
+
+
+
+
+
 ## [1.18.2](https://github.com/matteobruni/tsparticles/compare/react-tsparticles-demo@1.18.1...react-tsparticles-demo@1.18.2) (2021-07-31)
 
 **Note:** Version bump only for package react-tsparticles-demo