+++ demo/riot/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/matteobruni/tsparticles/compare/riot-particles-demo@1.7.2...riot-particles-demo@1.7.3) (2021-08-10)
+
+**Note:** Version bump only for package riot-particles-demo
+
+
+
+
+
 ## [1.7.2](https://github.com/matteobruni/tsparticles/compare/riot-particles-demo@1.7.1...riot-particles-demo@1.7.2) (2021-07-31)
 
 **Note:** Version bump only for package riot-particles-demo